requirements.txt
asgiref
dataclasses
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
djongo
dnspython
joblib
numpy
PyJWT
scikit-learn
scipy
sqlparse
threadpoolctl
typing_extensions
tzdata